this is partly prep for upcoming project
In this assignment, we are providing less direction and instruction for
what to do; this is to make the students show that they can handle these
higher-level tasks, and serves in part as preparation for the project,
where we *really* don't give them direction or
instructions.

In the Mastermind reading/activity, we mentioned that the class
diagramming was in part preparation for this; we should, in turn,
mention how this assignment fits with the upcoming project.